handle cloud syncing
